UIIN-1642: Fix impossibility to create `in transit` items report
